requirements.txt
pypdf2==1.25.1
Pillow==3.1.0
pylint==1.4.4
pyfiglet==0.7.4
